+++ apps/web/src/components/Theme/ThemeProvider/ThemeProviderContext.ts
@@ -3,12 +3,12 @@ import { createContext } from 'react'
 export type Thema = 'dark' | 'light'
 
 type ThemaContextType = {
-  getThema: () => Thema
+  theme: Thema
   toggleTheme: () => void
 }
 
-const ThemaProviderContext = createContext<ThemaContextType>({
-  getThema: () => 'dark',
+const ThemeProviderContext = createContext<ThemaContextType>({
+  theme: 'dark',
   toggleTheme: () => { }
 })
-export default ThemaProviderContext
+export default ThemeProviderContext
