+++ apps/web/src/components/Theme/ThemeProvider/index.ts
@@ -1,0 +1,2 @@
+export { ThemeProvider as default } from './ThemeProvider'
+export { default as ThemeProviderContext } from './ThemeProviderContext'
