--- apps/web/src/components/provider/ThemaProvider/index.ts
@@ -1,2 +1,0 @@
-export { ThemaProvider as default } from './ThemaProvider'
-export { default as ThemaProviderContext } from './ThemaProviderContext'
