--- apps/web/src/hooks/provider/Thema/useThema.ts
@@ -1,10 +1,0 @@
-import { useContext } from 'react'
-import ThemaProviderContext from '@provider/ThemaProvider/ThemaProviderContext'
-
-export default function useThema() {
-  const context = useContext(ThemaProviderContext)
-  if (!context) {
-    throw new Error('Wrong position')
-  }
-  return context
-}
